Initial sweep PR fixes - removing unneeded comments, renaming files, fixing britishisms
